# core-lending-supply-chain-finance-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-supply-chain-finance API
  description: Supply Chain Finance Management System (Factoring & Confirming)
  contact:
    name: Firefly Software Foundation
    url: https://getfirefly.io
    email: dev@getfirefly.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
...
paths:
...
  /api/v1/supply-chain-finance/counterparties/{id}:
    get:
      tags:
        - Counterparty
      summary: Get a counterparty by ID
      description: Retrieve a specific counterparty by its unique identifier
      operationId: getById_3
      parameters:
        - name: id
          in: path
          description: Unique identifier of the counterparty
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Counterparty found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CounterpartyDTO'
        '404':
          description: Counterparty not found
    put:
      tags:
        - Counterparty
      summary: Update an existing counterparty
      description: Update the details of an existing counterparty
      operationId: update_3
      parameters:
        - name: id
          in: path
          description: Unique identifier of the counterparty
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CounterpartyDTO'
        required: true
      responses:
        '200':
          description: Counterparty updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CounterpartyDTO'
        '400':
          description: Invalid counterparty data
        '404':
          description: Counterparty not found
    delete:
      tags:
        - Counterparty
      summary: Delete a counterparty
      description: Delete a counterparty from the system
      operationId: delete_3
      parameters:
        - name: id
          in: path
          description: Unique identifier of the counterparty
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Counterparty deleted successfully
        '404':
          description: Counterparty not found
...
  /api/v1/supply-chain-finance/advances/{id}:
    get:
      tags:
        - SupplyChainFinanceAdvance
      summary: Get a supply chain finance advance by ID
      description: Retrieve a specific advance by its unique identifier
      operationId: getById_5
      parameters:
        - name: id
          in: path
          description: Unique identifier of the advance
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Advance found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceAdvanceDTO'
        '404':
          description: Advance not found
    put:
      tags:
        - SupplyChainFinanceAdvance
      summary: Update an existing supply chain finance advance
      description: Update the details of an existing advance
      operationId: update_5
      parameters:
        - name: id
          in: path
          description: Unique identifier of the advance
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyChainFinanceAdvanceDTO'
        required: true
      responses:
        '200':
          description: Advance updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceAdvanceDTO'
        '400':
          description: Invalid advance data
        '404':
          description: Advance not found
    delete:
      tags:
        - SupplyChainFinanceAdvance
      summary: Delete a supply chain finance advance
      description: Delete an advance from the system
      operationId: delete_5
      parameters:
        - name: id
          in: path
          description: Unique identifier of the advance
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '204':
          description: Advance deleted successfully
        '404':
          description: Advance not found
...
  /api/v1/supply-chain-finance/counterparties:
    get:
      tags:
        - Counterparty
      summary: List/Search counterparties
      description: Retrieve a paginated list of counterparties based on filter criteria
      operationId: findAll_3
      parameters:
        - name: filterRequest
          in: query
          description: Filter criteria for counterparties
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestCounterpartyDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved counterparties
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
    post:
      tags:
        - Counterparty
      summary: Create a new counterparty
      description: Create a new counterparty in the system
      operationId: create_3
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CounterpartyDTO'
        required: true
      responses:
        '200':
          description: Counterparty created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CounterpartyDTO'
        '400':
          description: Invalid counterparty data
...
  /api/v1/supply-chain-finance/advances:
    get:
      tags:
        - SupplyChainFinanceAdvance
      summary: List/Search supply chain finance advances
      description: Retrieve a paginated list of advances based on filter criteria
      operationId: findAll_5
      parameters:
        - name: filterRequest
          in: query
          description: Filter criteria for advances
          required: true
          schema:
            $ref: '#/components/schemas/FilterRequestSupplyChainFinanceAdvanceDTO'
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved advances
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponse'
        '400':
          description: Invalid filter request
    post:
      tags:
        - SupplyChainFinanceAdvance
      summary: Create a new supply chain finance advance
      description: Create a new advance in the system
      operationId: create_5
      parameters:
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SupplyChainFinanceAdvanceDTO'
        required: true
      responses:
        '200':
          description: Advance created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/SupplyChainFinanceAdvanceDTO'
        '400':
          description: Invalid advance data
components:
  schemas:
...
    CounterpartyDTO:
      required:
        - agreementId
        - counterpartyName
        - financeType
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        agreementId:
          type: string
          format: uuid
        financeType:
          type: string
          enum:
            - FACTORING
            - CONFIRMING
        counterpartyName:
          maxLength: 255
          minLength: 0
          type: string
        taxId:
          maxLength: 50
          minLength: 0
          type: string
        registrationNumber:
          maxLength: 50
          minLength: 0
          type: string
        address:
          maxLength: 500
          minLength: 0
          type: string
        contactPerson:
          maxLength: 255
          minLength: 0
          type: string
        contactEmail:
          maxLength: 255
          minLength: 0
          type: string
        contactPhone:
          maxLength: 50
          minLength: 0
          type: string
        approved:
          type: boolean
        approvedLimit:
          type: number
        currentExposure:
          type: number
        creditRating:
          maxLength: 50
          minLength: 0
          type: string
        notes:
          maxLength: 2000
          minLength: 0
          type: string
...
    SupplyChainFinanceAdvanceDTO:
      required:
        - advanceAmount
        - advanceDate
        - currency
        - financeType
        - invoiceId
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        financeType:
          type: string
          enum:
            - FACTORING
            - CONFIRMING
        invoiceId:
          type: string
          format: uuid
        advanceDate:
          type: string
          format: date
        currency:
          type: string
          enum:
            - EUR
            - USD
            - GBP
            - CHF
            - JPY
            - CNY
        advanceAmount:
          type: number
        discountAmount:
          type: number
        feeAmount:
          type: number
        netAmount:
          type: number
        paymentReference:
          maxLength: 100
          minLength: 0
          type: string
        notes:
          maxLength: 2000
          minLength: 0
          type: string
    FilterOptions:
      type: object
      properties:
        caseInsensitiveStrings:
          type: boolean
        includeInheritedFields:
          type: boolean
      description: Filter options for customizing filter behavior
...
    PaginationRequest:
      type: object
      properties:
        pageNumber:
          type: integer
          description: The zero-based page number to retrieve.
          format: int32
          example: 0
          default: 0
        pageSize:
          type: integer
          description: The number of items per page.
          format: int32
          example: 10
          default: 10
        sortBy:
          type: string
          description: The field to sort the results by.
          example: name
        sortDirection:
          type: string
          description: The direction of sorting, either ASC or DESC.
          example: DESC
          default: DESC
      description: Represents a pagination request for retrieving paginated results, including page number, size, sort field, and direction.
    RangeFilter:
      type: object
      properties:
        ranges:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/RangeObject'
          description: Map of field names to their range criteria
      description: Represents range filters for numeric fields
    RangeObject:
      type: object
      properties:
        from:
          type: object
          description: Start value of the range
        to:
          type: object
          description: End value of the range
      description: Represents a range with from and to values
    PaginationResponse:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            type: object
            description: The list of items for the current page.
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    FilterRequestCounterpartyDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/CounterpartyDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination
...
    FilterRequestSupplyChainFinanceAdvanceDTO:
      required:
        - pagination
      type: object
      properties:
        filters:
          $ref: '#/components/schemas/SupplyChainFinanceAdvanceDTO'
        rangeFilters:
          $ref: '#/components/schemas/RangeFilter'
        pagination:
          $ref: '#/components/schemas/PaginationRequest'
        options:
          $ref: '#/components/schemas/FilterOptions'
      description: Generic filter request that includes both filter criteria and pagination